requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
uvloop>=0.19.0
sqlalchemy>=1.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
psutil>=5.8.0
python-dotenv>=0.19.0
cryptography>=3.4.0
python-multipart>=0.0.5
qrcode>=7.3.0

